requirements.txt
google-genai>=0.6.0
litellm>=1.40.0
google-adk>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=9.0.0
pandas>=2.0.0
openpyxl>=3.0.0
flask>=2.3.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
boto3>=1.34.0
monocle-apptrace>=0.1.0
pydantic>=2.0.0